+++ tests/measurement/test_measurement_control.py
@@ -14,14 +14,15 @@ from quantify.data.types import TUID
 from quantify.visualization.pyqt_plotmon import PlotMonitor_pyqt
 from quantify.visualization.instrument_monitor import InstrumentMonitor
 from quantify.utilities.experiment_helpers import load_settings_onto_instrument
-
+from tests.helpers import get_test_data_dir
+import tempfile
 try:
     from adaptive import SKOptLearner
     with_skoptlearner = True
 except ImportError:
     with_skoptlearner = False
 
-
+test_datadir = get_test_data_dir()
 
 
 def CosFunc(t, amplitude, frequency, phase):
@@ -158,16 +159,18 @@ class DummyHardwareGettable:
 class TestMeasurementControl:
     @classmethod
     def setup_class(cls):
+        cls.tmp_dir = tempfile.TemporaryDirectory()
+        dh.set_datadir(cls.tmp_dir.name)
         cls.MC = MeasurementControl(name="MC")
         # ensures the default datadir is used which is excluded from git
         cls.dummy_parabola = DummyParHolder("parabola")
-        dh.set_datadir(None)
 
     @classmethod
     def teardown_class(cls):
         cls.MC.close()
         cls.dummy_parabola.close()
-        dh.set_datadir(None)
+        dh._datadir = None
+        cls.tmp_dir.cleanup()
 
     def test_MeasurementControl_name(self):
         assert self.MC.name == "MC"
@@ -812,7 +815,7 @@ class TestMeasurementControl:
 
     def test_instrument_settings_from_disk(self):
         load_settings_onto_instrument(
-            self.dummy_parabola, TUID("20200814-134652-492-fbf254"), dh._test_dir
+            self.dummy_parabola, TUID("20200814-134652-492-fbf254"), test_datadir
         )
         assert self.dummy_parabola.x() == 40.0
         assert self.dummy_parabola.y() == 90.0
@@ -824,7 +827,7 @@ class TestMeasurementControl:
             ValueError, match='Instrument "the mac" not found in snapshot'
         ):
             load_settings_onto_instrument(
-                non_existing, TUID("20200814-134652-492-fbf254"), dh._test_dir
+                non_existing, TUID("20200814-134652-492-fbf254"), test_datadir
             )
 
         non_existing.close()